requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
configparser==5.0.1
dataclasses==0.6
docker-pycreds==0.4.0
filelock==3.0.12
future==0.18.2
gitdb==4.0.5
GitPython==3.1.11
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.11.21
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==0.17.0
numpy==1.19.4
packaging==20.4
pandas==1.1.4
pathtools==0.1.2
promise==2.3
protobuf==3.14.0
psutil==5.7.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.0
rfc3986==1.4.0
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.4
sentencepiece==0.1.91
sentry-sdk==0.19.4
shortuuid==1.0.1
six==1.15.0
sklearn==0.0
smmap==3.0.4
sniffio==1.2.0
subprocess32==3.5.4
threadpoolctl==2.1.0
tokenizers==0.9.3
tqdm==4.53.0
transformers==3.5.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wandb==0.10.11
watchdog==0.10.4
